be: seperate upload api with utils and fracture prediction api
