+++ be/app/api/__init__.py
@@ -1,3 +1,4 @@
 from .auth import router as auth
 from .student_chat import router as student_chat
 from .fracture_prediction import router as fracture_prediction
+from .upload import router as upload
